+++ dev/test/bulk-writer.ts
@@ -26,10 +26,18 @@ import {
   Timestamp,
   WriteResult,
 } from '../src';
-import {setTimeoutHandler} from '../src/backoff';
+import {
+  DEFAULT_BACKOFF_FACTOR,
+  DEFAULT_BACKOFF_INITIAL_DELAY_MS,
+  DEFAULT_BACKOFF_MAX_DELAY_MS,
+  MAX_RETRY_ATTEMPTS,
+  setTimeoutHandler,
+} from '../src/backoff';
 import {
   BulkWriterError,
-  DEFAULT_STARTING_MAXIMUM_OPS_PER_SECOND,
+  DEFAULT_INITIAL_OPS_PER_SECOND_LIMIT,
+  DEFAULT_JITTER_FACTOR,
+  DEFAULT_MAXIMUM_OPS_PER_SECOND_LIMIT,
 } from '../src/bulk-writer';
 import {
   ApiOverride,
@@ -44,7 +52,6 @@ import {
   updateMask,
   verifyInstance,
 } from './util/helpers';
-
 import api = proto.google.firestore.v1;
 
 // Change the argument to 'console.log' to enable debug output.
@@ -181,7 +188,6 @@ describe('BulkWriter', () => {
 
   afterEach(() => {
     verifyInstance(firestore);
-    expect(timeoutHandlerCounter).to.equal(0);
     setTimeoutHandler(setTimeout);
   });
 
@@ -255,7 +261,7 @@ describe('BulkWriter', () => {
       });
       expect(bulkWriter._rateLimiter.availableTokens).to.equal(100);
       expect(bulkWriter._rateLimiter.maximumCapacity).to.equal(
-        Number.POSITIVE_INFINITY
+        DEFAULT_MAXIMUM_OPS_PER_SECOND_LIMIT
       );
 
       bulkWriter = firestore.bulkWriter({
@@ -266,18 +272,18 @@ describe('BulkWriter', () => {
 
       bulkWriter = firestore.bulkWriter();
       expect(bulkWriter._rateLimiter.availableTokens).to.equal(
-        DEFAULT_STARTING_MAXIMUM_OPS_PER_SECOND
+        DEFAULT_INITIAL_OPS_PER_SECOND_LIMIT
       );
       expect(bulkWriter._rateLimiter.maximumCapacity).to.equal(
-        Number.POSITIVE_INFINITY
+        DEFAULT_MAXIMUM_OPS_PER_SECOND_LIMIT
       );
 
       bulkWriter = firestore.bulkWriter({throttling: true});
       expect(bulkWriter._rateLimiter.availableTokens).to.equal(
-        DEFAULT_STARTING_MAXIMUM_OPS_PER_SECOND
+        DEFAULT_INITIAL_OPS_PER_SECOND_LIMIT
       );
       expect(bulkWriter._rateLimiter.maximumCapacity).to.equal(
-        Number.POSITIVE_INFINITY
+        DEFAULT_MAXIMUM_OPS_PER_SECOND_LIMIT
       );
 
       bulkWriter = firestore.bulkWriter({throttling: false});
@@ -537,9 +543,9 @@ describe('BulkWriter', () => {
         ]),
         response: mergeResponses([
           successResponse(1),
-          failedResponse(Status.INTERNAL),
-          failedResponse(Status.INTERNAL),
-          failedResponse(Status.INTERNAL),
+          failedResponse(Status.CANCELLED),
+          failedResponse(Status.CANCELLED),
+          failedResponse(Status.CANCELLED),
         ]),
       },
       {
@@ -580,6 +586,7 @@ describe('BulkWriter', () => {
         'success',
       ]);
       expect(writeResults).to.deep.equal([1, 2, 3, 4]);
+      expect(timeoutHandlerCounter).to.equal(1);
     });
   });
 
@@ -605,6 +612,35 @@ describe('BulkWriter', () => {
     await bulkWriter.flush();
     expect(catchCalled).to.be.true;
     expect(onWriteErrorCalled).to.be.true;
+  });
+
+  it('retries INTERNAL errors for deletes', async () => {
+    const bulkWriter = await instantiateInstance([
+      {
+        request: createRequest([setOp('doc1', 'bar'), deleteOp('doc2')]),
+        response: mergeResponses([
+          failedResponse(Status.INTERNAL),
+          failedResponse(Status.INTERNAL),
+        ]),
+      },
+      {
+        request: createRequest([deleteOp('doc2')]),
+        response: successResponse(2),
+      },
+    ]);
+
+    let errorCaught = false;
+    bulkWriter
+      .set(firestore.doc('collectionId/doc1'), {
+        foo: 'bar',
+      })
+      .catch(() => {
+        errorCaught = true;
+      });
+    const del = bulkWriter.delete(firestore.doc('collectionId/doc2'));
+    await bulkWriter.close();
+    expect((await del).writeTime).to.deep.equal(new Timestamp(2, 0));
+    expect(errorCaught).to.be.true;
   });
 
   it('surfaces errors thrown by user-provided error callback', async () => {
@@ -679,6 +715,7 @@ describe('BulkWriter', () => {
       });
     await bulkWriter.close();
     expect(writeResult).to.equal(1);
+    expect(timeoutHandlerCounter).to.equal(3);
   });
 
   it('retries maintain correct write resolution ordering', async () => {
@@ -861,7 +898,7 @@ describe('BulkWriter', () => {
     return bulkWriter.close().then(() => verifyOpCount(2));
   });
 
-  it('retries individual writes that fail with ABORTED errors', async () => {
+  it('retries individual writes that fail with ABORTED and UNAVAILABLE errors', async () => {
     setTimeoutHandler(setImmediate);
     // Create mock responses that simulate one successful write followed by
     // failed responses.
@@ -999,7 +1036,16 @@ describe('BulkWriter', () => {
   });
 
   it('fails writes after all retry attempts failed', async () => {
-    setTimeoutHandler(setImmediate);
+    setTimeoutHandler((fn, timeout) => {
+      const expected =
+        DEFAULT_BACKOFF_INITIAL_DELAY_MS * Math.pow(1.5, timeoutHandlerCounter);
+      expect(timeout).to.be.within(
+        (1 - DEFAULT_JITTER_FACTOR) * expected,
+        (1 + DEFAULT_JITTER_FACTOR) * expected
+      );
+      timeoutHandlerCounter++;
+      fn();
+    });
     function instantiateInstance(): Promise<BulkWriter> {
       const overrides: ApiOverride = {
         batchWrite: () => {
@@ -1025,7 +1071,123 @@ describe('BulkWriter', () => {
       });
     return bulkWriter.close().then(() => {
       verifyOpCount(1);
+      expect(timeoutHandlerCounter).to.equal(MAX_RETRY_ATTEMPTS - 1);
     });
+  });
+
+  it('applies maximum backoff on retries for RESOURCE_EXHAUSTED', async () => {
+    setTimeoutHandler((fn, timeout) => {
+      timeoutHandlerCounter++;
+      expect(timeout).to.be.within(
+        (1 - DEFAULT_JITTER_FACTOR) * DEFAULT_BACKOFF_MAX_DELAY_MS,
+        (1 + DEFAULT_JITTER_FACTOR) * DEFAULT_BACKOFF_MAX_DELAY_MS
+      );
+      fn();
+    });
+    function instantiateInstance(): Promise<BulkWriter> {
+      const overrides: ApiOverride = {
+        batchWrite: () => {
+          const error = new GoogleError('Mock batchWrite failed in test');
+          error.code = Status.RESOURCE_EXHAUSTED;
+          throw error;
+        },
+      };
+      return createInstance(overrides).then(firestoreClient => {
+        firestore = firestoreClient;
+        return firestore.bulkWriter();
+      });
+    }
+    const bulkWriter = await instantiateInstance();
+    bulkWriter.onWriteError(err => err.failedAttempts < 5);
+    bulkWriter
+      .create(firestore.doc('collectionId/doc'), {
+        foo: 'bar',
+      })
+      .catch(err => {
+        expect(err instanceof BulkWriterError).to.be.true;
+        expect(err.code).to.equal(Status.RESOURCE_EXHAUSTED);
+        incrementOpCount();
+      });
+    return bulkWriter.close().then(() => {
+      verifyOpCount(1);
+      expect(timeoutHandlerCounter).to.equal(4);
+    });
+  });
+
+  it('uses the highest backoff found in the batch', async () => {
+    const expected = [
+      DEFAULT_BACKOFF_MAX_DELAY_MS,
+      DEFAULT_BACKOFF_INITIAL_DELAY_MS * DEFAULT_BACKOFF_FACTOR,
+    ];
+    setTimeoutHandler((fn, timeout) => {
+      // 1st batch should have max backoff. 2nd batch should have 1 round
+      // of backoff applied.
+      expect(timeout).to.be.within(
+        (1 - DEFAULT_JITTER_FACTOR) * expected[timeoutHandlerCounter],
+        (1 + DEFAULT_JITTER_FACTOR) * expected[timeoutHandlerCounter]
+      );
+      timeoutHandlerCounter++;
+      fn();
+    });
+    const bulkWriter = await instantiateInstance([
+      {
+        request: createRequest([createOp('doc1', 'bar'), setOp('doc2', 'bar')]),
+        response: mergeResponses([
+          failedResponse(Status.RESOURCE_EXHAUSTED),
+          failedResponse(Status.UNAVAILABLE),
+        ]),
+      },
+      {
+        request: createRequest([createOp('doc1', 'bar'), setOp('doc2', 'bar')]),
+        response: mergeResponses([
+          successResponse(1),
+          failedResponse(Status.UNAVAILABLE),
+        ]),
+      },
+      {
+        request: createRequest([setOp('doc2', 'bar')]),
+        response: successResponse(2),
+      },
+    ]);
+
+    bulkWriter.onWriteError(err => err.failedAttempts < 5);
+    bulkWriter.create(firestore.doc('collectionId/doc1'), {
+      foo: 'bar',
+    });
+    bulkWriter.set(firestore.doc('collectionId/doc2'), {
+      foo: 'bar',
+    });
+    return bulkWriter.close().then(() => {
+      expect(timeoutHandlerCounter).to.equal(2);
+    });
+  });
+
+  it('sends backoff batch after other enqueued batches', async () => {
+    setTimeoutHandler(setImmediate);
+    const bulkWriter = await instantiateInstance([
+      {
+        request: createRequest([createOp('doc1', 'bar')]),
+        response: failedResponse(Status.RESOURCE_EXHAUSTED),
+      },
+      {
+        request: createRequest([setOp('doc2', 'bar')]),
+        response: successResponse(1),
+      },
+      {
+        request: createRequest([createOp('doc1', 'bar')]),
+        response: successResponse(2),
+      },
+    ]);
+
+    bulkWriter.onWriteError(err => err.failedAttempts < 5);
+    bulkWriter.create(firestore.doc('collectionId/doc1'), {
+      foo: 'bar',
+    });
+    bulkWriter.flush();
+    bulkWriter.set(firestore.doc('collectionId/doc2'), {
+      foo: 'bar',
+    });
+    return bulkWriter.close();
   });
 
   describe('if bulkCommit() fails', async () => {